+++ apps/admin/components/lib/AntdSteps/index.tsx
@@ -1,0 +1,24 @@
+import { Steps } from 'antd';
+import React from 'react';
+
+const description = 'This is a description.';
+const App: React.FC = () => (
+  <Steps
+    direction="vertical"
+    size="small"
+    current={1}
+    items={[
+      { title: 'Finished', description },
+      {
+        title: 'In Progress',
+        description,
+      },
+      {
+        title: 'Waiting',
+        description,
+      },
+    ]}
+  />
+);
+
+export default App;
