--- apps/admin/containers/contentListComponent/index.module.scss
@@ -1,26 +1,0 @@
-@use 'styles/mixins' as m;
-
-.wrap {
-  .topBox {
-    @include m.flexbox(between);
-    margin-bottom: 40px;
-    .searchBox {
-      width: 520px;
-      @include m.flexbox(between);
-    }
-  }
-  .btnWarp {
-    @include m.flexbox(between);
-  }
-
-  .loadBtnWrap {
-    padding: 200px;
-    @include m.flexbox();
-  }
-
-  .btnBox {
-    width: 100%;
-    margin: 40px 0;
-    @include m.flexbox();
-  }
-}
